Support for setting OData related HTTP headers
